+++ revdep/checks/MorphoTools2/new/MorphoTools2.Rcheck/tests/testthat.R
@@ -1,7 +1,0 @@
-library(testthat)
-library(MorphoTools2)
-#library(visualTest)
-#library(png)
-
-test_check("MorphoTools2")
-